Make PlatformService#open(URL) work better
This change includes ij-platforms in the runtime classpath so that the
platform service can more effectively open a URL in a platform-specific
way.

We may want to consider removing the PlatformService API example from
this tutorial, since it complexifies the POM, but as things stand, it is
needed to open the URL most effectively on all platforms.