Add call-modern-from-legacy tutorial
This tutorial is intended to demonstrate usage of modern ImageJ features
from within a legacy ImageJ1 plugin.

For now there is only one example: generating and displaying a
spreadsheet with String fields. This is useful since the legacy
ResultsTable is limited to double values, whereas in ImageJ2 the Table
API supports whatever you want to throw at it.