call-modern-from-legacy: fix IJ2's plugin discovery
As pointed out by Curtis Rueden, Sezpoz will not be able to find the
plugins if it does not know about the appropriate class loader.

To tell Sezpoz about the class loader, we simply set the current
thread's context class loader to the class loader which was able to
load the plugin class, and we're set.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>