LoadAndDisplayDataset: fix showing of the dataset
It was previously assumed that the Swing UI was active and visible, such
that calling DisplayService#createDisplay would result in the new
display being automatically shown. But since the example is currently
written *not* to show any UI, we must explicitly tell the UI service to
show the dataset instead, by calling UIService#show(Object).

The alternative would be to instead call UIService#showUI() before
creating the display. The difference would be that the ImageJ main
window would then also be visible.