Use the new ImageJ convenience class
This updates the code to match the latest ImageJ2 codebase, which
includes the new and improved ImageJ convenience class.

As a consequence, the tutorial code has gotten shorter and hopefully
easier to understand. Just as importantly, the tutorial *POMs* have also
gotten shorter: they all include only a single ij-app dependency now,
which is much simpler than trying to understand exactly which ij-*
dependencies are appropriate in each situation.