Update the .factorypath files to match pom-scijava
We use a dirty little trick in .factorypath: we use the sezpoz .jar
that Eclipse's Maven plugin had to download to configure the Maven
projects properly. However, with the upgrade to a new SciJava POM,
the sezpoz version changed. As a result, the old .factorypath versions
did not work for users who never had an old version of the SciJava
POM and hence never downloaded sezpoz 1.9.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>