Update tutorials for 2.0.0-beta-7.1 release
This release moved some service framework classes into SciJava Common,
and changed how the IOService works. Tutorials have been updated
accordingly.

The only exception is the 'execute-commands' tutorial which for some
reason is broken with 2.0.0-beta-7.1. So I left it at 2.0.0-beta-7 for
now. We will need to fix it though, of course.