ExecuteCommands: flesh out javadoc
Most of this text was taken from an ImageJ2 documentation page:
    http://developer.imagej.net/writing-imagej2-plugins

That page was written before this tutorial existed, but this tutorial
covers many of the same material. So the relevant material from that
page has now been integrated into this tutorial, and removed from the
page itself in favor of a link.