Revamp how the "Hello World" command works
The first incarnation used the UIService, which meant the command could
not easily be used in a headless context. The new version has exactly
one input and exactly one output. It prompts for your name, then greets
you by storing the greeting as the output. The framework takes care of:

A) Harvesting the input value from the user
B) Displaying the output value to the user

Hopefully this symmetry is a simple yet effective illustration of
ImageJ2's typed command mechanism.