simple-commands: add ComputeStats example command
The simple-commands tutorial already had a very simple command
(HelloWorld) with a single string input and output, as well as a simple
GradientImage command with integer inputs and a dataset output. What was
missing was a command that takes a dataset as input but produces numbers
as output; the ComputeStats example now provides that.

These commands are useful for testing interoperability of image data
structures with ImageJ in both directions.