WidgetDemo: add double slider & scroll bar widgets
You can use doubles with sliders & scroll bar number widgets, but there
are limitations: the slider/scroll component will not respect fractional
min/maxes or step sizes. This is because ultimately the granularity of
the slider/scroll component is still hardcoded at the integer level.

What *is* now possible is to use the associated spinner component to set
a fractional value and the slider/scroll component will be set to the
truncated integer value without nuking the spinner's fractional value.