Specified swing interface for ROI display
originally when the add-ROI tutorial was written, the “swing” user
interface was the default. At some point after, the ImageJ 1 user
interface was set to a higher priority causing the UI to not be
visible. Specifying the UI as “swing” allows us to display the UI and
created ROI once more.

See also:
https://github.com/imagej/imagej-ui-swing/issues/16
https://github.com/imagej/imagej/issues/89