Add another example of Op creation
The "Random Blobs" op is another example of creating and running a new
op. This op takes desired blob size and number of blobs as parameters
and then creates an image using these specifications. The tutorial is
meant to show how ops can aid in image creation.

A main method has been included within the tutorial to allow the user to
output and view the image in the console window as ascii characters.