Added a tutorial class that uses "create a new op"
The “Random Blobs” class implements the further embellishes previously
made “create a new op” tutorial. Using a run method and Op.class, it
selects the number of random blobs needed as well as their size in a
main method and displays the image in ascii. The purpose of the
tutorial is to help further demonstrate the uses of the the Op class
aswell as the ImgLib library.