Draw method edited
In the “Random Blobs” op, ‘draw()’ was altered to use the inputted
radius and randomly generated center point to calculate the rest of the
neighborhood within each blob. To simplify the code, two methods were
created to evaluate whether or not all blob coordinates are within the
user inputted dimensions. If a coordinate is not within dimensions, it
is wrapped.

‘net.imglib2.util.Intervals’ was also imported so the utility method to
calculate the number of intervals within an image could be used.