Update tutorials to latest version of ImageJ
This bumps the parent to pom-imagej 7.1.0, which updates the version of
Javassist to 3.19.0-GA, fixing IJ1 patching problems when using Java 8.

This commit also makes the necessary code changes to account for
backwards incompatibilities in ImageJ OPS: specifically, the migration
of 'add' operations into the 'math' namespace.

Unfortunately, the update introduces other issues across the board.