Bump most parents to pom-imagej 15.8.0
We no longer need to declare Java 8, since it is now the default.

But we now must explicitly declare imagej-legacy and ij dependencies,
for those components which use it.

The only components not yet updated are add-two-datasets and using-ops,
because those tutorials need more serious work in order to update them.
So for now, they stay at 14.4.1.