Update parents to org.scijava:pom-scijava:12.0.0
See:
http://forum.imagej.net/t/split-boms-from-parent-configuration/2563

This update fills in the normal POM metadata in the aggregator POM,
but fills in dummy/placeholder metadata in each individual tutorial.
That way, people can copy/paste and replace the placeholders with
their own information, without ever accidentally advertising the
wrong authorship/organization/etc. metadata for their project.

This update also makes the necessary code changes to adjust for
backwards incompatibilities in the APIs.