Use a less ugly naming convention
This is a matter of opinion, but I much prefer dash-separated words.
I find it much easier to read when browsing the folder structure.

This also removes the sole ImgLib2+Python notebook, in favor of a
better ImageJ+Python example notebook coming in a couple of commits,
and cleans up the toplevel notebook's table of contents a little.